requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.5
sendgrid==6.10.0
starkbank-ecdsa==2.2.0
urllib3==2.5.0
